requirements.txt
PyQt6
gTTS
pygame
requests
SpeechRecognition
pyaudio
Pillow